chore: move cache key constant to index.ts
